fix rename OceanCompute to Compute
