+++ src/lib.ts
@@ -13,7 +13,7 @@ export * from './ddo/DDO'
 export * from './ddo/MetaData'
 
 export { CreateProgressStep } from './ocean/OceanAssets'
-export { ComputeJob, ComputeJobStatus } from './ocean/OceanCompute'
+export { ComputeJob, ComputeJobStatus } from './ocean/Compute'
 export { OrderProgressStep } from './ocean/utils/ServiceUtils'
 export {
     OceanPlatformTechStatus,
